fix: 프로필 페이지와 공개 프로필 페이지가 제대로 보이도록 페이지와 ROUTE 수정
