+++ frontend/src/api/axios/axios.custom.ts
@@ -1,7 +1,7 @@
-import instance from "@/api/axios/axios.instance";
 import { UserRegisterDataDto } from "@/types/dto/user.dto";
-import { GameMode, GameType } from "@/types/enum/game.enum";
 import { ChannelPolicy } from "@/types/enum/channel.enum";
+import { GameMode, GameType } from "@/types/enum/game.enum";
+import instance from "@/api/axios/axios.instance";
 
 const axiosMyInfoURL = "/user/me";
 export const axiosMyInfo = async (): Promise<any> => {
@@ -49,6 +49,21 @@ export const axiosGetUserProfileByNickname = async (
   try {
     const response = await instance.get(
       `${axiosGetUserProfileByNicknameURL}/${nickname}`,
+    );
+    return response;
+  } catch (error) {
+    console.log(error);
+    throw error;
+  }
+};
+
+const axiosGetUserGameRecordURL = "/game/record";
+export const axiosGetUserGameRecord = async (
+  nickname: string,
+): Promise<any> => {
+  try {
+    const response = await instance.get(
+      `${axiosGetUserGameRecordURL}/${nickname}`,
     );
     return response;
   } catch (error) {
