+++ frontend/src/ui/overview/profile/record.tsx
@@ -1,0 +1,35 @@
+import styled from "styled-components";
+import {
+  CellHeaderStyled,
+  CellStyled,
+  ChannelListContainerStyled,
+  RowStyled,
+  TableBody,
+  TableHeader,
+} from "@/ui/overview/game/game-list";
+
+export default function Record({ games }: { games: any }) {
+  return (
+    <ChannelListContainerStyled>
+      <TableHeader>
+        <CellHeaderStyled>Player</CellHeaderStyled>
+        <CellHeaderStyled>Result</CellHeaderStyled>
+        <CellHeaderStyled>Type</CellHeaderStyled>
+        <CellHeaderStyled>Mode</CellHeaderStyled>
+        <CellHeaderStyled>Date</CellHeaderStyled>
+      </TableHeader>
+      <TableBody>
+        {games &&
+          games.map((game: any, idx: number) => (
+            <RowStyled key={idx} onClick={() => {}} className="channel">
+              <CellStyled className="align-center">{game.player}</CellStyled>
+              <CellStyled className="align-center">{game.result}</CellStyled>
+              <CellStyled className="align-center">{game.type}</CellStyled>
+              <CellStyled className="align-center">{game.mode}</CellStyled>
+              <CellStyled className="align-center">{game.date}</CellStyled>
+            </RowStyled>
+          ))}
+      </TableBody>
+    </ChannelListContainerStyled>
+  );
+}
